Fix Swiftlint warnings (file length) in files related to simulator tunnel
